block: move testing utilities in separate module
